+++ apps/native-component-list/src/screens/UI/DateTimePickerScreen.tsx
@@ -1,0 +1,89 @@
+import { DateTimePicker, DatePickerProps } from '@expo/ui/components/DatePicker';
+import { Picker } from '@expo/ui/components/Picker';
+import * as React from 'react';
+import { Platform, ScrollView, Text, View } from 'react-native';
+
+import { Page, Section } from '../../components/Page';
+
+export default function DatePickerScreen() {
+  const [selectedDate, setSelectedDate] = React.useState(new Date());
+
+  const displayOptions =
+    Platform.OS === 'ios' ? ['compact', 'graphical', 'wheel'] : ['picker', 'input'];
+  const [selectedIndex, setSelectedIndex] = React.useState(0);
+
+  const typeOptions = ['date', 'hourAndMinute', 'dateAndTime'];
+  const [typeIndex, setTypeIndex] = React.useState(0);
+
+  function printPickerType() {
+    const str = displayOptions[selectedIndex];
+    return `${str.charAt(0).toUpperCase()}${str.slice(1)} picker`;
+  }
+
+  return (
+    <ScrollView>
+      <Page>
+        <Section title="Selected Date">
+          <Text>{selectedDate.toDateString()}</Text>
+        </Section>
+        <Section title="Selected Time">
+          <Text>{selectedDate.toTimeString()}</Text>
+        </Section>
+        <Section title={printPickerType()}>
+          <View style={{ gap: 20 }}>
+            <View
+              onLayout={(event) => {
+                console.log('📏 DatePickerView resized:', event.nativeEvent.layout);
+              }}>
+              <DateTimePicker
+                title=""
+                onDateSelected={(date) => {
+                  setSelectedDate(date);
+                }}
+                displayedComponents={
+                  typeOptions[typeIndex] as DatePickerProps['displayedComponents']
+                }
+                initialDate={selectedDate.toISOString()}
+                iosVariant={displayOptions[selectedIndex] as DatePickerProps['iosVariant']}
+                androidVariant={displayOptions[selectedIndex] as DatePickerProps['androidVariant']}
+                style={{ height: Platform.select({ android: 520, ios: undefined }) }}
+                showVariantToggle
+                is24Hour
+              />
+            </View>
+
+            <View
+              style={{ gap: 20 }}
+              onLayout={(event) => {
+                console.log('📏 Picker View resized:', event.nativeEvent.layout);
+              }}>
+              <Picker
+                options={displayOptions}
+                selectedIndex={selectedIndex}
+                onOptionSelected={({ nativeEvent: { index } }) => {
+                  setSelectedIndex(index);
+                }}
+                variant="segmented"
+                style={{ height: 30 }}
+              />
+
+              <Picker
+                options={typeOptions}
+                selectedIndex={typeIndex}
+                onOptionSelected={({ nativeEvent: { index } }) => {
+                  setTypeIndex(index);
+                }}
+                variant="segmented"
+                style={{ height: 30 }}
+              />
+            </View>
+          </View>
+        </Section>
+      </Page>
+    </ScrollView>
+  );
+}
+
+DatePickerScreen.navigationOptions = {
+  title: 'DatePicker',
+};