+++ packages/expo-ui/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### 🎉 New features
 
+- Adds `DateTimePicker` component.
+
 ### 🐛 Bug fixes
 
 ### 💡 Others